Added file cleanup back into project to removed old downloads.

+++ DABApp/Droid/DabApi/FileManagement.cs
@@ -34,8 +34,17 @@ namespace DABApp.Droid
             {
                 var doc = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData);
                 var fileName = Path.Combine(doc, $"{episodeId}.{extension}");
-                File.Delete(fileName);
+                if (File.Exists(fileName))
+                {
+                    Debug.WriteLine($"Deleted episode {episodeId}");
+                    File.Delete(fileName);
+                }
+                else
+                {
+                    Debug.WriteLine($"Ignored deletion of episode {episodeId}");
+                }
                 return true;
+
             }
             catch (Exception e)
             {
