+++ DABApp/iOS/DabApi/FileManagement.cs
@@ -65,8 +65,14 @@ namespace DABApp.iOS
             { 
                 var doc = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData);
                 var fileName = Path.Combine(doc, $"{episodeId}.{extension}");
+                if (File.Exists(fileName))
+                { 
                 Debug.WriteLine($"Deleted episode {episodeId}");
                 File.Delete(fileName);
+                } else
+                {
+                    Debug.WriteLine($"Ignored deletion of episode {episodeId}");
+                }
                 return true;
             }
             catch (Exception e)
